# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulp API",
    "description": "Pulp social-listening HTTP API. This spec is generated from the Rust backend and is the authoritative API contract.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/backfill": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "trigger_backfill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Single-channel backfill complete"
          },
          "202": {
            "description": "All-channel backfill accepted"
          }
        }
      }
    },
    "/api/admin/collect/{channel}": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "trigger_collect",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "description": "Channel name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection run triggered"
          }
        }
      }
    },
...
    "/api/channels/{channel}": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "`GET /api/channels/:channel` — fetch a single channel config by name.\nReturns 404 if the channel has never been configured.",
        "operationId": "getChannel",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "description": "Channel name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelConfig"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "channels"
        ],
        "summary": "`PUT /api/channels/:channel` — create or update a channel config. Body fields\nare all optional: `enabled`, `poll_interval` (seconds, default 900), and\n`credentials` (channel-specific JSON). For Reddit the only (optional) credential\nis `{ \"user_agent\" }` — Reddit's public search needs no client id/secret. For\nGitHub the credentials are a `{ \"token\", ... }` filter object. When `enabled` is\nomitted the channel's current enabled state is preserved. When `credentials` is\nomitted the previously-stored credentials are preserved (not wiped) — pass an\nexplicit `credentials` object to change them.",
        "operationId": "upsert",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "description": "Channel name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelConfig"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/{channel}/backfill": {
      "post": {
        "tags": [
          "channels"
        ],
        "operationId": "backfill",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "description": "Channel name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/{channel}/cleanup": {
      "post": {
        "tags": [
          "channels"
        ],
        "operationId": "cleanup",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "description": "Channel name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanupBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/config/ai": {
      "get": {
        "tags": [
          "config"
        ],
        "operationId": "get_ai_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "operationId": "update_ai_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid settings (e.g. enabled without base_url/model)"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiConfigUpdate": {
        "type": "object",
        "description": "Partial update. Every field is optional; omitted (or null) fields are left\nunchanged. For `api_key`, an empty string clears the stored key.",
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "AiConfigView": {
        "type": "object",
        "description": "AI-filter settings as exposed to clients. The API key is never returned;\n`api_key_set` reports only whether one is configured.",
        "required": [
          "enabled",
          "base_url",
          "model",
          "api_key_set"
        ],
        "properties": {
          "api_key_set": {
            "type": "boolean"
          },
          "base_url": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "model": {
            "type": "string"
          }
        }
      },
...
      "BackfillBody": {
        "type": "object",
        "required": [
          "days"
        ],
        "properties": {
          "days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "BackfillRequest": {
        "type": "object",
        "required": [
          "since"
        ],
        "properties": {
          "channel": {
            "type": [
              "string",
              "null"
            ]
          },
          "since": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BackfillResult": {
        "type": "object",
        "description": "Result of a channel backfill: a human-readable status message.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ChannelBody": {
        "type": "object",
        "properties": {
          "credentials": {},
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "poll_interval": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "ChannelConfig": {
        "type": "object",
        "required": [
          "channel",
          "enabled",
          "credentials",
          "poll_interval",
          "updated_at",
          "max_backfill_days"
        ],
        "properties": {
          "caught_up_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The caught-up watermark: how far this channel is contiguously\ningested. Distinct from `last_polled_at`, which is merely the last\nattempt time."
          },
          "channel": {
            "type": "string"
          },
          "credentials": {},
          "enabled": {
            "type": "boolean"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_polled_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "max_backfill_days": {
            "type": "integer",
            "format": "int64"
          },
          "poll_interval": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CleanupBody": {
        "type": "object",
        "required": [
          "dry_run"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean"
          }
        }
      },
      "CleanupPreview": {
        "type": "object",
        "description": "Preview returned by a dry-run cleanup: how many mentions would be removed,\nplus up to 10 samples.",
        "required": [
          "count",
          "sample"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "sample": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MentionSample"
            }
          }
        }
      },
      "CleanupResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CleanupPreview"
          },
          {
            "$ref": "#/components/schemas/CleanupResult"
          }
        ],
        "description": "A cleanup call returns a preview on `dry_run` and a delete count otherwise.\nUntagged so the wire shape is exactly one of the two inner objects."
      },
      "CleanupResult": {
        "type": "object",
        "description": "Result of a real (non-dry-run) cleanup: how many mentions were deleted.",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "MentionSample": {
        "type": "object",
        "required": [
          "id",
          "title",
          "url"
        ],
        "properties": {
          "author": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "repo": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}